fix #13 PPS format sample playback is broken.

+++ src/fmgen/types.h
@@ -34,8 +34,8 @@ typedef signed int int32;
 
 	typedef unsigned char	BYTE;
 	typedef unsigned short	WORD;
-	typedef unsigned long	DWORD;
-	
+//	typedef unsigned long	DWORD;
+	typedef unsigned 		DWORD;
 
 
 #endif
