+++ src/pmdwin/pcmmusdriver.h
@@ -18,7 +18,8 @@ typedef unsigned int uint;
 typedef unsigned long ulong;
 typedef unsigned char BYTE;
 typedef unsigned short WORD;
-typedef unsigned long DWORD;
+//typedef unsigned long DWORD;
+typedef unsigned DWORD;
 #ifndef __cplusplus
 typedef unsigned char bool;
 #endif
